+++ c/triplet.h
@@ -51,13 +51,13 @@
 /* in the input. */
 long tpl_get_triplets_reciprocal_mesh_at_q(long *map_triplets,
                                            long *map_q,
-                                           int (*grid_address)[3],
+                                           long (*grid_address)[3],
                                            const long grid_point,
-                                           const int mesh[3],
-                                           const int is_time_reversal,
+                                           const long mesh[3],
+                                           const long is_time_reversal,
                                            const long num_rot,
-                                           TPLCONST int (*rotations)[3][3],
-                                           const int swappable);
+                                           TPLCONST long (*rotations)[3][3],
+                                           const long swappable);
 /* Irreducible grid-point-triplets in BZ are stored. */
 /* triplets are recovered from grid_point and triplet_weights. */
 /* BZ boundary is considered in this recovery. Therefore grid addresses */
@@ -66,28 +66,28 @@ long tpl_get_triplets_reciprocal_mesh_at_q(long *map_triplets,
 /* Number of ir-triplets is returned. */
 long tpl_get_BZ_triplets_at_q(long (*triplets)[3],
                               const long grid_point,
-                              TPLCONST int (*bz_grid_address)[3],
+                              TPLCONST long (*bz_grid_address)[3],
                               const long *bz_map,
                               const long *map_triplets,
                               const long num_map_triplets,
-                              const int mesh[3]);
+                              const long mesh[3]);
 void tpl_get_integration_weight(double *iw,
                                 char *iw_zero,
                                 const double *frequency_points,
                                 const long num_band0,
-                                TPLCONST int relative_grid_address[24][4][3],
-                                const int mesh[3],
+                                TPLCONST long relative_grid_address[24][4][3],
+                                const long mesh[3],
                                 TPLCONST long (*triplets)[3],
                                 const long num_triplets,
-                                TPLCONST int (*bz_grid_address)[3],
+                                TPLCONST long (*bz_grid_address)[3],
                                 const long *bz_map,
                                 const double *frequencies1,
                                 const long num_band1,
                                 const double *frequencies2,
                                 const long num_band2,
                                 const long tp_type,
-                                const int openmp_per_triplets,
-                                const int openmp_per_bands);
+                                const long openmp_per_triplets,
+                                const long openmp_per_bands);
 void tpl_get_integration_weight_with_sigma(double *iw,
                                            char *iw_zero,
                                            const double sigma,
@@ -100,10 +100,10 @@ void tpl_get_integration_weight_with_sigma(double *iw,
                                            const long num_band,
                                            const long tp_type);
 
-int tpl_is_N(const long triplet[3], const int *grid_address);
+long tpl_is_N(const long triplet[3], const long *grid_address);
 void tpl_set_relative_grid_address(
-  int tp_relative_grid_address[2][24][4][3],
-  TPLCONST int relative_grid_address[24][4][3],
+  long tp_relative_grid_address[2][24][4][3],
+  TPLCONST long relative_grid_address[24][4][3],
   const long tp_type);
 
 #endif
